requirements.txt
PyQt6
moviepy
pandas
playwright
numpy
elevenlabs
google-genai
pyautogui
